requirements.txt
asgiref==3.4.1
astroid==2.9.2
attrs==21.4.0
autopep8==1.6.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
Django==4.0.1
django-phonenumber-field==6.0.0
djangorestframework==3.13.1
gunicorn==20.1.0
h11==0.12.0
iniconfig==1.1.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
packaging==21.3
phonenumbers==8.12.40
platformdirs==2.4.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pylint==2.12.2
pyparsing==3.0.6
pytest==6.2.5
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
uvicorn==0.16.0
wrapt==1.13.3
